requirements.txt
Django==1.8
ecdsa==0.13
Fabric==1.10.1
ftfy==4.0.0
gnureadline==6.3.3
httplib2==0.9.1
ipython==3.1.0
paramiko==1.15.2
psycopg2==2.6
pycrypto==2.6.1
python-dateutil==2.4.2
pytz==2015.2
requests==2.6.0
six==1.9.0
twilio==3.8.0
